# Add `wp_get_theme_data_template_parts`: create public API to access `templateParts` from `theme.json`