--- a/src/wp-includes/global-styles-and-settings.php
+++ b/src/wp-includes/global-styles-and-settings.php
@@ -448,7 +448,26 @@ function wp_get_theme_directory_pattern_slugs() {
  * return string[]
  */
 function wp_get_theme_template_part_metadata() {
-	return WP_Theme_JSON_Resolver::get_theme_data( array(), array( 'with_supports' => false ) )->get_template_parts();
+	$cache_group    = 'theme_json';
+	$cache_key      = 'wp_get_theme_template_part_metadata';
+	$can_use_cached = ! wp_is_development_mode( 'theme' );
+
+	$metadata = false;
+	if ( $can_use_cached ) {
+		$metadata = wp_cache_get( $cache_key, $cache_group );
+		if ( false !== $metadata ) {
+			return $metadata;
+		}
+	}
+
+	if ( false === $metadata ) {
+		$metadata = WP_Theme_JSON_Resolver::get_theme_data( array(), array( 'with_supports' => false ) )->get_template_parts();
+		if ( $can_use_cached ) {
+			wp_cache_set( $cache_key, $metadata, $cache_group );
+		}
+	}
+
+	return $metadata;
 }
 
 /**