@oandregal 2023-08-07
This is what the [perf tests](https://github.com/WordPress/wordpress-develop/actions/runs/5786249582/job/15680582758?pr=4971) report:

| Base line | This PR? |
| --- | --- |
| <img width="455" alt="Captura de ecrã 2023-08-07, às 17 30 58" src="https://github.com/WordPress/wordpress-develop/assets/583546/a916b003-08e1-4254-8f98-9a9bab012fa2"> | <img width="385" alt="Captura de ecrã 2023-08-07, às 17 28 23" src="https://github.com/WordPress/wordpress-develop/assets/583546/0a648247-c9ef-4984-97a4-e70afa33cacb"> |

In the past, I was told the PR's perf job was not reporting data for the own PR, so I am not sure I should trust these numbers? @felixarntz Should I still run my benchmark locally?