@joemcgill 2023-08-07
@oandregal the benchmarks reported in the screenshots that you've pasted are comparing WP 6.1.1 ([the baseline](https://github.com/WordPress/wordpress-develop/blob/trunk/.github/workflows/performance.yml#L45)) with your specific commit, so it's not measuring the impact of this specific change. This limitation is something we've got [a ticket to address](https://core.trac.wordpress.org/ticket/58358), but have not done so yet, so taking an individual benchmark if you can would be helpful. Thanks!