@oandregal 2023-08-08
@felixarntz @joemcgill this is ready: created the trac ticket, provided testing instructions and a performance report.

Performance-wise, this has less impact than I originally thought. It is still valuable, though. Specially, because that the main reason for this PR is to introduce a public method to retrieve data from `theme.json` (consumers have to use private APIs so far).