@oandregal 2023-08-08
Side-notes about performance report:

- It is still costly to prepare one. Would you think https://core.trac.wordpress.org/ticket/58358 could be prioritized? It's a bottleneck to analyze PRs manually. As a result, we do less pre-merge performance analysis. For example, in this PR, where it was not strictly necessary, I've spent a few hours.

- I'm surprised about the differences depending on the method. Not the numbers themselves, but the ratio for block vs classic themes. [Codevitals](https://www.codevitals.run/project/wordpress) reports server time as almost equal for classic and blocks while the other methods report classic themes being 2x faster than blocks (`curl` is more pessimistic and lies in 2.3x with `benchmark-*` tools being more optimistic around 1.9x).