@felixarntz 2023-08-09
@oandregal 
> I'm surprised about the differences depending on the method. Not the numbers themselves, but the ratio for block vs classic themes. [Codevitals](https://www.codevitals.run/project/wordpress) reports server time as almost equal for classic and blocks while the other methods report classic themes being 2x faster than blocks (`curl` is more pessimistic and lies in 2.3x with `benchmark-*` tools being more optimistic around 1.9x).

I have no idea, but that's definitely something we should research more. I wonder whether there's a notable difference in how the two environments set up the test sites? Otherwise, there's a good chance it is because the automated workflow used for Codevitals uses a headless browser, while `benchmark-server-timing` does not (similar to `curl`). I wonder if the headless browser setup adds some overhead that skews the data somehow - or vice versa?