@oandregal 2023-08-09
🤔 I just realized the [test data](https://github.com/WordPress/wordpress-develop/blob/trunk/.github/workflows/performance.yml#L136) in use by CI is different from the “dummy” content that comes by default. I'd need to use that dataset to know whether it affected these tests.