@oandregal 2023-08-09
> I'm surprised about the differences depending on the method.

I just tried `compare-wp-performance` by @swissspidy and this is the result of running `./run.sh latest trunk`.

`wp-total (p50)` for classic themes is at 20.88ms while it sits at 31.64ms for block themes, a 1.5x (unlike the ~1x of codevitals and ~2x of `curl` and `benchmark-*` tools). We have a slight disagreement on the numbers.