@felixarntz 2023-08-09
@oandregal Related to this change, @swissspidy shared some relevant feedback on how creating more and more individual wrapper functions for `theme.json` data is not exactly a sustainable approach, see [Slack message](https://wordpress.slack.com/archives/C02KGN5K076/p1691507507013079). Wouldn't it be more reasonable to consider implementing caching directly into `WP_Theme_JSON_Resolver`, or implement a wrapper class that comes with caching?