@oandregal 2023-08-10
> @oandregal Related to this change, @swissspidy shared some relevant feedback on how creating more and more individual wrapper functions for theme.json data is not exactly a sustainable approach, see [Slack message](https://wordpress.slack.com/archives/C02KGN5K076/p1691507507013079). Wouldn't it be more reasonable to consider implementing caching directly into WP_Theme_JSON_Resolver, or implement a wrapper class that comes with caching?

Happy to re-evaluate caching, though, we need the individual functions: they are the public API. Given the discussion around caching doesn't modify the API, I'm going to commit this to continue follow-up work. Please, let the conversation continue.

So, caching: the resolver already [caches data](https://github.com/WordPress/wordpress-develop/blob/trunk/src/wp-includes/class-wp-theme-json-resolver.php#L668). The issue is that it has to invalidate some of it once per request, resulting in some expensive processes having to run twice. I'm looking into changing how it works to make it run once. Take `get_core_data` as an example: the next image show all its callees to the right. They are called twice when they should have been called once:

<img width="1440" alt="Captura de ecrã 2023-08-10, às 12 07 05" src="https://github.com/WordPress/wordpress-develop/assets/583546/831cf13a-8c8b-4ada-83b8-e6af5dc2e667">

By introducing the public APIs (and caching "at the edge" when it makes sense, such as for settings, styles, templateParts, etc.) I'm hoping to be able to decouple some of the internal processes that theme.json classes do, removing the unnecessary ones depending on the data being accessed, making sure caches are use effectively, etc.

Hope this helps?