@oandregal 2023-08-10
> A class can be a public API too, we don‘t have to add functions for everything.

I know. Though, note that `WP_Theme_JSON` and `WP_Theme_JSON_Resolver` classes are indeed private, which means we cannot ask the consumers to use them for accessing data (normal use case).

> My concern was mostly that I‘m worried we‘ll end up creating X new functions each with its own caching wrapper around that class, leading to poor API design and lots of repetition

I understand, those are valid concerns I also have. For reference, this is [the plan](https://github.com/WordPress/gutenberg/issues/45171). Note that the plan is not to have a micro-method for everything. The plan is to have five methods. One per [top-level keys that can exists in a `theme.json`](https://developer.wordpress.org/block-editor/how-to-guides/themes/theme-json/):

```json
{
  "settings": {},
  "styles": {},
  "templateParts": {},
  "customTemplates": {},
  "patterns": {}
}
```

And the public API:

- `wp_get_global_settings()`. Shipped in 5.9.0.
- `wp_get_global_styles()`. Shipped in 5.9.0.
- `wp_get_theme_directory_pattern_slugs`. Shipped in 6.3.
- `wp_get_theme_data_template_parts()`. Added by this PR.
- `wp_get_theme_data_custom_templates()`. TBD.

The rationale for this is that each piece has its own requirements (translated or not translated, how sanitization works, etc.). By having these methods, we can improve how things work by refactoring the private classes/processes as needed without breaking backwards compatibility.

Does this overview alleviate some of your concerns?